==> 2. pre_SQL_preprocessing.ipynb <==
## Because to implement data into a table in SQL, I need to create the table beforehand and define the columns and their types, in order to enable myself to implement data into such a table, I perform preliminary data cleaning - I correct the data types and remove unnecessary columns. Additionally, I change the column names from the automatically obtained ones to the target names to facilitate further work.

import numpy as np
import pandas as pd

df = pd.read_csv('C:/Users/Maciej/Documents/projekt_studia/real_estate/final_result.csv')

### First glance at the dataframe and its basic information

df.head()

df.columns

df.info()

#### check for the % of the missing values (np.nan) in every column

df.isna().sum() / len(df) * 100


### A more in-depth look at the values of the columns and the deletion of those I found useless for future analysis.

to_drop = []

for i in df.columns:
    if df[i].isna().sum() / len(df) * 100 > 95:
        to_drop.append(i)

An instant deletion of every column that has percentage of the missing values above 95%

to_drop.append('Unnamed: 0')

Column 'Unnamed: 0' has been generated automatically as an index and is deleted to remove the redundancy

df['Rodzaj umowy'].unique()
to_drop.append('Rodzaj umowy')

If there are two discrete values in a column, and one of them is np.nan, I assume that the missing values should be replaced with the only remaining value. Therefore, a column that has the same value for every record is useless for analysis and should be deleted.

#### I will now analyze following column in terms of a price therefore I need to firstly update datatype of a price column to int

df[df['Cena'] == 'Zarezerwowano'] = np.nan
df.dropna(subset='Cena', inplace=True)
df['Cena'] = df['Cena'].str.replace('zł', '')
df['Cena'] = df['Cena'].str.replace(' ', '').astype(int)

df['Cena/metr'] = df['Cena/metr'].str.replace('zł/m²', '')
df['Cena/metr'] = df['Cena/metr'].str.replace(' ', '').astype(int)

to_drop.append('Inwestycja')
df['Inwestycja'].unique()

print(df['Forma własności'].unique())
print('średnia cena mieszkania dla zbioru:',df[df['Forma własności'] == 'Własność']['Cena'].mean())
print('średnia cena (forma własnpści = "spóldzielcze..."', df[df['Forma własności'] == 'Spółdzielcze własnościowe']['Cena'].mean())
to_drop.append('Forma własności')

df['Liczba poziomów mieszkania'].unique()

print(df['Typ kuchni'].unique())
for i in df['Typ kuchni'].unique():
    print('średnia cena dla:',i, df[df['Typ kuchni'] == i]['Cena'].mean())
print(len(df[df['Typ kuchni'] == 'Brak']))
to_drop.append('Typ kuchni')

print(df['Liczba łazienek'].unique())

print(df[df['Liczba łazienek'] > 1]['Cena'].mean())
print(len(df[df['Liczba łazienek'] > 1]))

print(df['łazienka razem z WC'].unique())
to_drop.append('łazienka razem z WC')

print(df['Taras'].unique()) 

df['Stan nieruchomości'].unique()

print(df['Typ budynku'].unique())

print(df['Materiał budowlany'].unique())

print(df['Aktualizacja'].unique())
to_drop.append('Aktualizacja')

print(df['Numer ogłoszenia'].unique())
to_drop.append('Numer ogłoszenia')

print(df['Liczba Punktów Podbić'].unique())

print(df['Liczba pięter'].unique())

print(df['Liczba sypialni'].unique())

print(df['Balkon'].unique())

df['Ogrzewanie'].unique()

print(df['Stolarka okienna'].unique())
to_drop.append('Stolarka okienna')

print(df['Dostępne od'].unique())
to_drop.append('Dostępne od')

to_drop.append('Pow. ogródka')

to_drop.append('Liczba odsłon')
to_drop.append('Ogrzewanie')

to_drop

#### List of all the columns to drop from the dataset

df.drop(columns=to_drop, inplace=True)
df.columns

#### Run-through the columns that have remained in dataset

len(df.columns)

df.info()

### In the next part - I change the datatype of the columns and I fill null values if there is such need. I divided dataset into 3 partitions for easier inspection of dataframe.

df.iloc[:5, :9]

df.iloc[:5, 9:18]

df.iloc[:5, 18:]

most_frequent_building = df['Typ budynku'].value_counts().keys()[0]
most_frequent_material = df['Materiał budowlany'].value_counts().keys()[0]
most_frequent_floors = df['Liczba pięter'].value_counts().keys()[0]

Preparation of variables containing most frequent values of the columns that are suitable this type of fill

df['Pow. całkowita'] = df['Pow. całkowita'].str.replace(' m²', '')
df['Pow. całkowita'] = df['Pow. całkowita'].str.replace(',', '.')
df['Data dodania'] = df['Data dodania'].str.replace('.', '-')
df['Liczba pięter'] = df['Liczba pięter'].str.replace('/', most_frequent_floors)

Adjustments in the way values are stored in the columns to avoid conflicts with the planned data type

df.loc[df['Stan nieruchomości'] == 'DO WYKOŃCZENIA', 'Stan nieruchomości'] = 'Do wykończenia'

df['Liczba poziomów mieszkania'].fillna(1, inplace=True)
df['Liczba łazienek'].fillna(1, inplace=True)
df['Taras'].fillna('Nie', inplace=True)
df['Typ budynku'].fillna(most_frequent_building, inplace=True)
df['Materiał budowlany'].fillna(most_frequent_material, inplace=True)
df['Liczba pięter'].fillna(most_frequent_floors, inplace=True)
df['Liczba sypialni'].fillna(1, inplace=True)
df['Balkon'].fillna('Nie', inplace=True)

Filling missing values with the most frequent values or most logical value for the flat.

df.isna().sum() / len(df)

After filling missing values the only columns to have missing values are the street and the status of the property.

df['id'] = df['id'].astype('int')
df['Pokoje'] = df['Pokoje'].astype('int')
df['Pow. całkowita'] = df['Pow. całkowita'].astype('float')
df['Liczba poziomów mieszkania'] = df['Liczba poziomów mieszkania'].astype('int')
df['Liczba łazienek'] = df['Liczba łazienek'].astype('int')
df['Rok budowy'] = df['Rok budowy'].astype('int')
df['Data dodania'] = pd.to_datetime(df['Data dodania'], format='%d-%m-%Y')
df['Data dodania'] = df['Data dodania'].dt.strftime('%Y-%m-%d')
df['Data dodania'] = pd.to_datetime(df['Data dodania'], format='%Y-%m-%d')
df['Liczba Punktów Podbić'] = df['Liczba Punktów Podbić'].astype('int')
df['Piętro'] = df['Piętro'].astype('int')
df['Liczba pięter'] = df['Liczba pięter'].astype('int')
df['Liczba sypialni'] = df['Liczba sypialni'].astype('int')

Change of the datatype for future analysis and SQL database implementation

rename_dict = {'Cena' : 'price',
              'Cena/metr' : 'price/m',
              'Pokoje' : 'rooms',
              'Miasto' : 'city',
              'Dzielnica' : 'district',
              'Ulica' : 'street',
              'Stan nieruchomości' : 'status',
              'Pow. całkowita' : 'meters',
              'Liczba poziomów mieszkania' : 'apartment_floors',
              'Liczba łazienek' : 'bathroom',
              'Taras' : 'terrace',
              'Rynek' : 'market',
              'Typ budynku' : 'building',
              'Materiał budowlany' : 'material',
              'Rok budowy' : 'year',
              'Data dodania' : 'adv_date',
              'Liczba Punktów Podbić' : 'upvotes',
              'Piętro' : 'floor',
              'Liczba pięter' : 'building_floors',
              'Liczba sypialni' : 'bedroom',
              'Balkon' : 'balcony'}

df.rename(columns=rename_dict, inplace=True)

Variable name change for easier future operations

df.info()

### Saving results to csv that is going to be implemented to the PostgreSQL table

df.to_csv('final_result_updated.csv')
